move and rename src/util.ts to components/utils_sasCode.ts
