+++ client/src/components/notebook/Controller.ts
@@ -4,7 +4,7 @@
 import * as vscode from "vscode";
 import { getSession } from "../../connection";
 import { Deferred, deferred } from "../utils";
-import { wrapCode } from "../../util";
+import { wrapCode } from "../utils_sasCode";
 
 export class NotebookController {
   readonly controllerId = "sas-notebook-controller-id";
